% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanWifiData.R
\name{CleanWifiData}
\alias{CleanWifiData}
\title{Cleans variables in wifi data}
\usage{
CleanWifiData(wifi.data)
}
\arguments{
\item{wifi.data}{A wifi dataset containing variables named "NetworkType",
"Channel", "Latitude", "Longitude", "Maxspeed", "Minspeed", "Firstdetection",
and "LastDetection"}
}
\value{
A dataframe of the same size with cleaned wifi data
}
\description{
Changes appropriate variables to factors and corrects time-logging problem
(hours incorrectly specified as minutes and minutes incorrectly specified as seconds)
}
\examples{

}
\references{

}
