% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertGPSCoords.R
\name{ConvertGPSCoords}
\alias{ConvertGPSCoords}
\title{GPS coordinate converter}
\usage{
ConvertGPSCoords(gps.data)
}
\arguments{
\item{gps.data}{a gps dataframe with a UTC column}
}
\value{
The original dataframe with longitude and latitude coordinates added
}
\description{
Converts GPS x and y coordinates to longitude and latitude based
on linear model
}
\examples{

}
\references{

}
