% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertGPSTimes.R
\name{ConvertGPSTimes}
\alias{ConvertGPSTimes}
\title{GPS times converter}
\usage{
ConvertGPSTimes(gps.data, gps.time.offset)
}
\arguments{
\item{gps.data}{a gps dataframe with a UTC column}

\item{gps.time.offset}{a chron times object that specifies how many hours the
gps time data is off from the wifi time data
(gps.time.offset = gps time - wifi time)}
}
\value{
A dataframe identical to the gps dataframe with the corrected time
(gps.data$time)
}
\description{
Function to convert GPS UTC to correct chron times vector
}
\examples{

}
\references{

}
