% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvalPass.R
\name{EvalPass}
\alias{EvalPass}
\title{Evaluate passengers according to cutoff values}
\usage{
EvalPass(obs, stops.data, phones, dist.tol, speed.tol, time.tol, rate.tol,
  pass.mod)
}
\arguments{
\item{obs}{A single observation from a wifi dataset that is to be classified
as a passenger or not.}

\item{stops.data}{a matching list of stops with longitude and latitude data}

\item{phones}{vector of phone manufacturers}

\item{dist.tol}{chron times object that determines how close to a stop a device
must be to be labelled a passenger}

\item{speed.tol}{speed in meters per second where observations with speed
greater than this will be classified as non-passengers}

\item{time.tol}{the number of seconds where observations that are
detected for less than this time will be classified as non-passengers}

\item{rate.tol}{maximum transfer rate tolerance in kbps?}

\item{pass.mod}{A random forest model to classify "unknown" manufacturers as
valid manufacturers or not.}
}
\value{
The same observation from a wifi dataset with a "passenger" variable
that indicates whether or not it was determined to be a bus passenger
(1: passenger, 0: not a passenger). If the observation is determined to be a
passenger the first and last stop (first.stop and last.stop respectively) are
added as new columns to the observation.
}
\description{
A function to calculate whether or not a single observation is a passenger
based on specified cutoff values
}
\examples{

}
\references{

}
