% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportODFlows.R
\name{ExportODFlows}
\alias{ExportODFlows}
\title{Export origin-destination flows}
\usage{
ExportODFlows(labelled.data, stops.data, folder.name = "Output",
  o.d.flows.name = "O-DFlows.csv",
  compact.name = "O-DFlowsCompact.csv",
  aggregate.name = "O-DFlowsAggregate.csv", remove.cols = TRUE,
  rename.cols = TRUE)
}
\arguments{
\item{labelled.data}{A wifi dataset with passenger labels (from the
`PredictLabels` function)}

\item{stops.data}{a matching list of stops with longitude and latitude data}

\item{folder.name}{The name of the folder to create (if it doesn't already
exist) where the exported o-d flows will be stored}

\item{o.d.flows.name}{The name of csv which the o-d flows will be stored in}

\item{compact.name}{The name of the csv which the compact o-d flows will be
stored in}

\item{aggregate.name}{The name of the csv which the aggregated o-d flows will
be stored in}

\item{remove.cols}{Removes irrelevant columns}

\item{rename.cols}{Renames (capitalizes) columns in labelled data}
}
\value{

}
\description{
Exports o-d flows to a new folder with a full, compact, and aggregated
version.
}
\examples{

}
\references{

}
