% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictLabels.R
\name{PredictLabels}
\alias{PredictLabels}
\title{Detection/Prediction for Bus Passengers}
\usage{
PredictLabels(wifi.data, gps.data, stops.data, phones, dist.tol, speed.tol,
  time.tol, rate.tol = 6000, gps.time.offset)
}
\arguments{
\item{wifi.data}{A wifi dataset containing variables named "NetworkType",
"Channel", "Latitude", "Longitude", "Maxspeed", "Minspeed", "Firstdetection",
and "LastDetection"}

\item{gps.data}{a gps dataframe with a UTC column}

\item{stops.data}{a matching list of stops with longitude and latitude data}

\item{phones}{vector of phone manufacturers}

\item{dist.tol}{chron times object that determines how close to a stop a device
must be to be labelled a passenger}

\item{speed.tol}{speed in meters per second where observations with speed
greater than this will be classified as non-passengers}

\item{time.tol}{the number of seconds where observations that are
detected for less than this time will be classified as non-passengers}

\item{rate.tol}{maximum transfer rate tolerance in kbps?}

\item{gps.time.offset}{a chron times object that specifies how many hours the
gps time data is off from the wifi time data
(gps.time.offset = gps time - wifi time)}
}
\value{
PredictLabels returns the wifi dataset with labelled passengers. For
each detected passenger, the first and last stops are included as separate
columns (first.stop and last.stop respectively)
}
\description{
Predicts labels of passengers based on proximity (time) to ground-truth
stops, manufacturer, minimum speed, detection time and maximum transfer rate
}
\examples{

}
\references{

}
