% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SyncTimes.R
\name{SyncTimes}
\alias{SyncTimes}
\title{Synchronize Bus Stop Arrival Times}
\usage{
SyncTimes(gps.data, stops.data, gps.time.offset,
  chon.time.tol = chron::times("0:05:00"), dist.tol = 0.001,
  first.cutoff = chron::times("0:05:00"),
  last.cutoff = chron::times("0:05:00"))
}
\arguments{
\item{gps.data}{a gps dataframe with a UTC column}

\item{stops.data}{a matching list of stops with longitude and latitude data}

\item{gps.time.offset}{a chron times object that specifies how many hours the
gps time data is off from the wifi time data
(gps.time.offset = gps time - wifi time)}

\item{chon.time.tol}{Chron times object which is the tolerance for times that are
determined to be on the same part of the route}

\item{dist.tol}{Radius of circle used to determine closeness of points}

\item{first.cutoff}{Chron times object where the first stop must occur within
this cutoff of the beginning of the route}

\item{last.cutoff}{Chron times object where the last stop must occur within
this cutoff of the end of the route}
}
\value{
A modified stops.data dataframe with the estimated time that the bus reached each stop
}
\description{
Synchronize the time a bus reaches each stop based on topological principals
}
\examples{

}
\references{

}
